feat: add create nodes feature to editor
Owners/authors/editors should now be able to create nodes
from the editor interface.

Needed to update a few packaged and add a module declaration
so TypeScript would be appeased.

Additionally, added some comments with TODO's related to
increasing the robustness of the application to avoid
headaches down the line.