feat: show add asset button only on edit node
Can't add assets to nodes that do not exist,
so only show "Add Asset" button when we're on an "edit"
route of an existing node.

This makes implementation of mutations simpler in my head,
as we do not have to deal with the case when an author
is creating a node and wants to immediately add assets to it.
There would be no node_id to create the asset with!
Then we would need to work out how to save the appropriate types
and data of the asset locally, then properly chain the
creation of the node, and the the creation and addition of assets
to it afterwards.

Not enough time to contemplate this implementation detail!