feat: add update assets and asset content
Can now update assets and add content to text assets.
In the future when creating new widgets, will need
to use really convuluted value update logic
due to the nested nature of nodeData when received.

Couldn't think of another reasonable way to isolate asset
state such that it could be contained with the widget
and have the update called from the Editor component
without using some odd/hack-y work around.

Additionally need to resolve an error in the cope-client-utils
package. a typo in the assetUpdate function results in the
asset update not working. Will make a PR to fix that soon