feat: add simple widgets for other asset types
Add Image and Video input widget, which are just text inputs
for links to existing files on the (e.g. youtube share links
or links to images hosted on another server)

Added additional types and input widgets to map of widgets to
asset types. This is tentative!!! More work should be done on
building on these widgets once we add additional functionality
on the backend.