fix: change index to start at 1, vs 0
Encountering weird edge cases with updating indexes appropriately
when index for first element was set to 0. Specfically,
if we have 3 assets on a node, and delete the first, we end up
with assets being index at [1, 1] as opposed to [0, 1] after
the delete.

Modifying how we create and update assets so everything starts
at 1 addressed these edge cases and led to expected results,
but if we want to start indexing at 0 further exploration into
why this is happening is needed.